Edit README.md
The content written on the README.md is initial as the commit is made
before the project is made. The skills to be acquired after finishing
the project is subject to changes throughout its development. Another
edit on README.md will be made while working on the project and after
the project is successfully deployed. A section explaining the
reflection that has been made while working on the project will be
subsequently added.